fix(api): change files

+++ api/src/user/user.service.ts
@@ -1,19 +1,67 @@
+/* eslint-disable @typescript-eslint/no-unsafe-return */
+/* eslint-disable @typescript-eslint/no-unsafe-member-access */
+/* eslint-disable @typescript-eslint/no-unsafe-argument */
+/* eslint-disable @typescript-eslint/no-misused-promises */
+/* eslint-disable @typescript-eslint/no-unsafe-assignment */
+/* eslint-disable @typescript-eslint/no-floating-promises */
+
 import {
+  BadRequestException,
   ForbiddenException,
   Injectable,
   NotFoundException,
 } from '@nestjs/common';
+
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { StrapiService } from '../strapi/strapi.service';
+import { CustomerStripe } from '../stripe/classes/customer.stripe';
 
 @Injectable()
 export class UserService {
-  constructor(private readonly strapiService: StrapiService) {}
+  constructor(
+    private readonly strapiService: StrapiService,
+    private readonly stripeCustomer: CustomerStripe,
+  ) {}
 
-  create(createUserDto: CreateUserDto) {
+  async create(createUserDto: CreateUserDto) {
     try {
-      return this.strapiService.postData('users', createUserDto);
+      const strapiUser = (await this.strapiService.postData(
+        'users',
+        createUserDto,
+      )) as CreateUserDto;
+
+      if (!strapiUser || !strapiUser.username) {
+        throw new BadRequestException('Failed to create user in Strapi');
+      }
+
+      const stripeCustomer = await this.stripeCustomer.createCustomer(
+        strapiUser.email,
+        strapiUser.username,
+      );
+
+      if (!stripeCustomer || !stripeCustomer.id) {
+        throw new BadRequestException('Failed to create customer in Stripe');
+      }
+
+      const updatedStrapiUser: {
+        username: string;
+        email: string;
+        id: number;
+        stripe_customer_id?: string;
+      } = await this.strapiService.updateData(`users/${strapiUser.id}`, {
+        stripe_customer_id: stripeCustomer.id,
+      });
+
+      if (!updatedStrapiUser || !updatedStrapiUser.stripe_customer_id) {
+        throw new BadRequestException(
+          'Failed to update Strapi user with Stripe ID',
+        );
+      }
+
+      console.log('Strapi user updated with Stripe ID:', updatedStrapiUser);
+
+      return updatedStrapiUser;
     } catch (error) {
       console.error('Error creating user:', error);
       throw new ForbiddenException(
@@ -40,6 +88,15 @@ export class UserService {
     }
   }
 
+  findOneByEmail(email: string) {
+    try {
+      return this.strapiService.getDataByField('users', 'email', email);
+    } catch (error) {
+      console.error('Error fetching user by email:', error);
+      throw new NotFoundException(`User with email ${email} not found`);
+    }
+  }
+
   update(id: number, updateUserDto: UpdateUserDto) {
     try {
       return this.strapiService.updateData(`users/${id}`, updateUserDto);
@@ -51,14 +108,37 @@ export class UserService {
     }
   }
 
-  remove(id: number) {
+  remove = async (id: number) => {
     try {
-      return this.strapiService.deleteData(`users/${id}`);
+      const user = await this.strapiService.getDataById('users', id);
+
+      if (!user) {
+        throw new NotFoundException(`User with ID ${id} not found`);
+      }
+
+      const deletedStripeCustomer = await this.stripeCustomer.deleteCustomer(
+        user.stripe_customer_id,
+      );
+
+      if (!deletedStripeCustomer) {
+        throw new BadRequestException('Failed to delete customer in Stripe');
+      }
+
+      const strapiCustomer = await this.strapiService.deleteData(`users/${id}`);
+
+      if (!strapiCustomer) {
+        throw new BadRequestException('Failed to delete user in Strapi');
+      }
+
+      return {
+        deletedStripeCustomer,
+        strapiCustomer,
+      };
     } catch (error) {
       console.error('Error deleting user:', error);
       throw new ForbiddenException(
         `You do not have permission to delete user with ID ${id}`,
       );
     }
-  }
+  };
 }
